requirements.txt
Pillow==10.1.0
python-docx==1.0.1
python-dotenv==1.0.0
